Support use of EC2 instance profiles for authentication
When using an instance profile, the access key ID & secret key are passed
along automatically.

Authentication to ECR will be now be attempted iff any of the AWS related
parameters are set. This means that if you're using instance profiles and
*not* needing to switch roles then you'll have to pass in at least the
AWS region parameter for now to force ECR authentication.